Chia Health Checker for client v1.1.7 - PI, Linux, Mac ~BETA~
This is for testing purposes. A user submitted this to me to help out the MAC, Linux and PI clients. Adding files here so you can review the code without download.

PLEASE REVIEW AND USE AT YOUR OWN RISK - I quickly reviewed the code and saw nothing malicious. Please review and confirm as well.

Also if modifications are needed for mac/linux please feel free to make them and submit to me by either a commit suggestion on git hub or message me on reddit.

All credit for this goes to reddit user SMC0881

#!/bin/sh
#Tested on Ubuntu running on Raspberry Pi4
#My version of Ubuntu for Pi did not have netstat installed. Run sudo apt install net-tools.
#This should work for Mac and regular Linux versions too.
#